Allow patching the downloaded kernels
Older kernels often cannot be compiled on newer systems for various
reasons. Some of the problems have a fix in the upstream, so we apply
selected commits (if they can be applied).

This applies dfbd199a7cfe ("selinux: use kernel linux/socket.h for
genheaders and mdp") which prevents errors of kind:

    error: #error New address family defined, please update secclass_map.